(REF) Move BaseCommand::parseOptionalOption() to OptionalOption::parse()
